refactor: split postgres_repository.go into focused modules
Split the 1518-line postgres_repository.go into four focused files:
- postgres_repository.go (150 lines): struct, constructor, transaction helpers, utility functions
- position_log_write.go (545 lines): Create, Update, outbox variants, insert helpers
- position_log_read.go (710 lines): FindByID, FindByAccountID, List, query helpers, batch loaders
- position_log_batch.go (150 lines): CreateBatch, ID mapping

All files remain in the same package with identical behavior.
